+++ src/changelog.md
@@ -1,3 +1,2 @@
 ## changelog
-- add potential support for serial devices (ex: `/dev/ttyUSB*`)
-- add read-only support for troublesome inverters (set `"TroubleMode": "yes"` in `appsettings.json`)
+- improve command execution retry strategy